+++ contracts/periphery/contracts/s2s/examples/RemarkDemo.sol
@@ -69,7 +69,7 @@ contract RemarkDemo is SmartChainXApp, Ownable {
         );
     }
 
-    function onMessageDelivered(bytes4 lane, uint256 nonce, bool result) external override {
+    function onMessageDelivered(bytes4 lane, uint64 nonce, bool result) external override {
         // TODO: add your code here if you want to do something after the message delivered
     }
 }
